I've moved all of the container files into the containers folder. This is to improve the readability and understanding for future devs.
